mudanças no main e maquina de estados pros objetos
