+++ src/statemachine.h
@@ -1,21 +1,25 @@
 #include "raylib.h"
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
 
 typedef enum GameScreen {MENU, GAMEPLAY, ENDING} GameScreen;
 typedef enum PuzzleObjectState {ORIGINAL, ALTERED} PuzzleObjectState;
 typedef enum PuzzleObjectType {INTERACTIVE, COLECTABLE} PuzzleObjectType;
 typedef enum StateEventStatus {STATE_CHANGED, STATE_UNCHANGED} StateEventStatus;
 
-typedef struct PuzzleObject{
-  PuzzleObjectState state;
-  PuzzleObjectType type;
+// Object structure
+typedef struct GameObject{
+    PuzzleObjectState state;
+    PuzzleObjectType type;
 
-  const char* name;
-  Texture2D image;
-  Rectangle bounds;
-  Vector2 position;
-  Vector2 size;
-
-}PuzzleObject;
+    int id;
+    const char* name;
+    Texture2D texture;
+    Rectangle bounds;
+    Vector2 position;
+    Vector2 size;
+} GameObject;
 
 typedef struct StateEvent { 
   StateEventStatus status;
@@ -23,21 +27,26 @@ typedef struct StateEvent {
 
 }StateEvent;
 
-PuzzleObject armario();
-PuzzleObject caixa1;
-PuzzleObject caixa2;
-PuzzleObject buraco;
-PuzzleObject diario;
-PuzzleObject vitrola;
-PuzzleObject garrafa;
-PuzzleObject estante;
-PuzzleObject mapa;
-PuzzleObject pirata;
-PuzzleObject varaPesca;
-PuzzleObject chave;
-PuzzleObject bilhete;
-PuzzleObject dente;
+GameObject* objects;
+int objectCount;
 
-StateEvent processEvent(PuzzleObject utility, PuzzleObject target);
+GameObject armario();
+GameObject caixa1;
+GameObject caixa2;
+GameObject buraco;
+GameObject diario;
+GameObject vitrola;
+GameObject garrafa;
+GameObject estante;
+GameObject mapa;
+GameObject pirata;
+GameObject varaPesca;
+GameObject chave;
+GameObject bilhete;
+GameObject dente;
+
+StateEvent processEvent(GameObject utility, GameObject target);
 
 void CheckMouseHover(void);
+
+GameObject* SetupObjects(void);
